producers/requirements.txt
kafka-python==2.0.2
google-analytics-data==0.18.12
python-dotenv==1.0.0
requests==2.31.0
